+++ packages/kysely/test/forms/get.spec.ts
@@ -1,7 +1,5 @@
 import { describe, expect, it } from "vitest";
 import { getTestServer } from "../helpers/test-server";
-import { FormType } from "@authhero/adapter-interfaces";
-import { FormFieldType } from "@authhero/adapter-interfaces";
 
 describe("forms", () => {
   describe("get", () => {
@@ -18,16 +16,6 @@ describe("forms", () => {
 
       const { id } = await data.forms.create("tenantId", {
         name: "Basic Form",
-        type: FormType.CUSTOM,
-        fields: [
-          {
-            id: "field1",
-            type: FormFieldType.TEXT,
-            label: "Name",
-            required: true,
-            name: "field1",
-          },
-        ],
       });
 
       const form = await data.forms.get("tenantId", id);
@@ -35,18 +23,6 @@ describe("forms", () => {
       expect(form).toMatchObject({
         id,
         name: "Basic Form",
-        fields: [
-          {
-            id: "field1",
-            type: "text",
-            label: "Name",
-            required: true,
-          },
-        ],
-        controls: [],
-        layout: {
-          columns: 1,
-        },
       });
     });